Detect masked fields from Confluent Connect
Confluent has updated Connecto REST API to return asterisks (or unicode
dots !?) on the responses as a security measure. This has resulted in
application using the REST API for automation breaking in some ways.

In the case gafkalo this means we can't detect if a connector secret
needs to be updated or not and we must push this every time.

This commit also changes the Plan output to show only changed fields